Rearrange code into smaller files
There is now separate files for `Queue`, `Sequence`, `Step`, `Execution.
The Zipper for each structure has been moved to its correspondent file.

There are still more improvements to be made organizing the code, but
for now I think this is good enough to keep moving forward.
